backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Vector Database (pgvector)
pgvector==0.2.5

# Google Gemini AI
google-generativeai==0.3.2

# Document Processing
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
markdown==3.5.2
python-magic==0.4.27
reportlab==4.0.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation and Settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Utilities
python-dotenv==1.0.1
redis==5.0.1

# Text Processing
tiktoken==0.5.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Code Quality
ruff==0.1.14
black==24.1.1
mypy==1.8.0
